Fix breaking password reset
