+++ routes/login/forget-password.js
@@ -2,6 +2,7 @@ const sendEmail = require('../helpers').email
 const { DB } = include('config/')
 const { datastructures } = include('routes/helpers/')
 const jwt = require('jsonwebtoken');
+const { isPasswordSecure } = require('./password-requirement')
 
  // Function to send password reset email
 async function sendResetEmail(email, html) {
@@ -71,7 +72,6 @@ function verifyTokenFields(decoded, res) {
 exports.getResetToken = async (req, res, next) => {
   const { token } = req.params;
   req.session.errormessage = '';
-
   jwt.verify(token, process.env.APP_SECRET, async function(err, decoded) {
     if(decoded) {
       if (!verifyTokenFields(decoded, res)) {
@@ -140,33 +140,3 @@ exports.updatePassword = async (req, res, next) => {
   });
 };
 
-
-exports.isPasswordSecure = (password) => {
-  // Check complexity (contains at least one uppercase, lowercase, number, and special character)
-  const uppercaseRegex = /[A-Z]/;
-  const lowercaseRegex = /[a-z]/;
-  const numberRegex = /[0-9]/;
-  const specialCharRegex = /[!@#$%^&*\(\)]/;
-  // Check against common passwords (optional)
-  const commonPasswords = ['password', '123456', 'qwerty'];
-  const checkPass = {
-    'pw-length': !(password.length < 8),  // Check length
-    'pw-upper': uppercaseRegex.test(password),
-    'pw-lower': lowercaseRegex.test(password),
-    'pw-number': numberRegex.test(password),
-    'pw-special': specialCharRegex.test(password),
-    'pw-common': !commonPasswords.includes(password),
-  };
-
-  const msgs = {
-    'pw-length': 'Password is too short!',
-    'pw-upper': 'Password requires at least one uppercase letter!',
-    'pw-lower': 'Password requires at least one lowercase letter!',
-    'pw-number': 'Password requires at least one numberal!',
-    'pw-special': 'Password requires at least one of the special characters: !@#$%^&*()',
-    'pw-common': 'Password cannot be a commonly used password!',
-  };
-
-  return Object.keys(checkPass).filter((key) => !checkPass[key]).map((key) => msgs[key]).join('\n');
-
-}
